[REFACTOR] Move NotificationToaster to shared component folder
